requirements.txt
absl-py==1.4.0
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Deprecated==1.2.14
Flask==2.3.2
flatbuffers==23.5.26
frozenlist==1.4.0
future==0.18.3
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.57.0
h5py==3.9.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keras==2.13.1
libclang==16.0.6
line-bot-sdk==3.2.0
Markdown==3.4.4
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.1.1
pydantic_core==2.4.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
Werkzeug==2.3.6
wrapt==1.15.0
yarl==1.9.2
